Refactored HostedDatabaseMigrator to Common namespace
